+++ ubo_app/services/090-assistant/ubo_handle.py
@@ -17,9 +17,24 @@ async def setup(register_reducer: ReducerRegistrar) -> None:
     await init_service()
 
 
+def binary_env_provider() -> dict[str, str]:
+    from ubo_app.constants import (
+        GOOGLE_CLOUD_SERVICE_ACCOUNT_KEY_SECRET_ID,
+        OPENAI_API_KEY_SECRET_ID,
+    )
+
+    return {
+        'GOOGLE_CLOUD_SERVICE_ACCOUNT_KEY_SECRET_ID': (
+            GOOGLE_CLOUD_SERVICE_ACCOUNT_KEY_SECRET_ID
+        ),
+        'OPENAI_API_KEY_SECRET_ID': OPENAI_API_KEY_SECRET_ID,
+    }
+
+
 register(
     service_id='assistant',
     label='Assistant',
     setup=setup,
     binary_path='bin/ubo-assistant',
+    binary_env_provider=binary_env_provider,
 )